Extract portfolio workflow contracts
