+++ src/app/services/portfolio_source_readiness.py
@@ -1,6 +1,6 @@
 from typing import Any
 
-from app.contracts.portfolio import (
+from app.contracts.portfolio_workflow import (
     PortfolioReadinessBucket,
     PortfolioReadinessIndicator,
     PortfolioReadinessReason,
